+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 * Avoid warnings by not using `var` in ripe-sdk-demo .js files
+* Set renderer output encoding to `sRGBEncoding`
 
 ### Fixed